# apps/backend/docs/openapi/openapi.json
{
  "components" : {
    "schemas" : {
      "PageResponsePostSummaryResponse" : {
        "description" : "오프셋 페이지네이션 응답",
        "properties" : {
          "content" : {
            "description" : "페이지 항목 목록",
            "items" : {
              "$ref" : "#/components/schemas/PostSummaryResponse"
            },
            "type" : "array"
          },
          "page" : {
            "description" : "0-based 페이지 번호",
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "description" : "페이지 크기",
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "description" : "전체 항목 수",
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "description" : "전체 페이지 수",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "page", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
      "PostCreateRequest" : {
        "description" : "게시글 작성 요청",
        "properties" : {
          "author" : {
            "description" : "작성자",
            "maxLength" : 20,
            "minLength" : 1,
            "type" : "string"
          },
          "content" : {
            "description" : "본문",
            "maxLength" : 10000,
            "minLength" : 1,
            "type" : "string"
          },
          "title" : {
            "description" : "제목",
            "maxLength" : 200,
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "author", "content", "title" ],
        "type" : "object"
      },
      "PostResponse" : {
        "description" : "게시글 상세 응답",
        "properties" : {
          "author" : {
            "description" : "작성자",
            "type" : "string"
          },
          "content" : {
            "description" : "본문",
            "type" : "string"
          },
          "createdAt" : {
            "description" : "작성 시각",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "게시글 ID",
            "format" : "uuid",
            "type" : "string"
          },
          "title" : {
            "description" : "제목",
            "type" : "string"
          },
          "updatedAt" : {
            "description" : "수정 시각",
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "author", "content", "createdAt", "id", "title", "updatedAt" ],
        "type" : "object"
      },
      "PostSummaryResponse" : {
        "description" : "게시글 목록 항목 응답",
        "properties" : {
          "author" : {
            "description" : "작성자",
            "type" : "string"
          },
          "createdAt" : {
            "description" : "작성 시각",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "게시글 ID",
            "format" : "uuid",
            "type" : "string"
          },
          "title" : {
            "description" : "제목",
            "type" : "string"
          }
        },
        "required" : [ "author", "createdAt", "id", "title" ],
        "type" : "object"
      },
      "PostUpdateRequest" : {
        "description" : "게시글 수정 요청",
        "properties" : {
          "content" : {
            "description" : "본문",
            "maxLength" : 10000,
            "minLength" : 1,
            "type" : "string"
          },
          "title" : {
            "description" : "제목",
            "maxLength" : 200,
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "content", "title" ],
        "type" : "object"
      },
      "ProblemDetail" : {
        "description" : "RFC 9457 ProblemDetail 에러 응답",
        "properties" : {
          "code" : {
            "description" : "기계 분기용 안정 코드",
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "items" : {
              "properties" : {
                "field" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              },
              "required" : [ "field", "message" ],
              "type" : "object"
            },
            "type" : "array"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "traceId" : {
            "description" : "관측용 트레이스 ID",
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "code", "status", "title" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "간단한 게시판 CRUD API",
    "title" : "Board API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/posts" : {
      "get" : {
        "operationId" : "listPosts",
        "parameters" : [ {
          "description" : "0-based 페이지 번호",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "페이지 크기(1-100)",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePostSummaryResponse"
                }
              }
            },
            "description" : "게시글 페이지"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "잘못된 요청"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "서버 오류"
          }
        },
        "summary" : "활성 게시글을 최신순으로 페이지 조회한다",
        "tags" : [ "board-controller" ]
      },
      "post" : {
        "operationId" : "createPost",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostResponse"
                }
              }
            },
            "description" : "작성된 게시글"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "잘못된 요청"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "서버 오류"
          }
        },
        "summary" : "게시글을 작성한다",
        "tags" : [ "board-controller" ]
      }
    },
    "/api/v1/posts/{id}" : {
      "delete" : {
        "operationId" : "deletePost",
        "parameters" : [ {
          "description" : "게시글 ID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "삭제됨"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "잘못된 요청"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "리소스를 찾을 수 없음"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "서버 오류"
          }
        },
        "summary" : "게시글을 소프트삭제한다",
        "tags" : [ "board-controller" ]
      },
      "get" : {
        "operationId" : "getPost",
        "parameters" : [ {
          "description" : "게시글 ID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostResponse"
                }
              }
            },
            "description" : "게시글 상세"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "잘못된 요청"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "리소스를 찾을 수 없음"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "서버 오류"
          }
        },
        "summary" : "게시글 상세를 조회한다",
        "tags" : [ "board-controller" ]
      },
      "put" : {
        "operationId" : "updatePost",
        "parameters" : [ {
          "description" : "게시글 ID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostResponse"
                }
              }
            },
            "description" : "수정된 게시글"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "잘못된 요청"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "리소스를 찾을 수 없음"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "서버 오류"
          }
        },
        "summary" : "게시글을 수정한다",
        "tags" : [ "board-controller" ]
      }
    }
  },
  "servers" : [ {
    "url" : "/"
  } ]
}